some reorganization have no code except `getSuspenseCache` create a new SuspenseCache any more
