+++ src/react/cache/getSuspenseCache.ts
@@ -1,5 +1,5 @@
-import type { SuspenseCacheOptions } from '../cache/index.js';
-import { SuspenseCache } from '../cache/index.js';
+import type { SuspenseCacheOptions } from './index.js';
+import { SuspenseCache } from './SuspenseCache.js';
 import type { ApolloClient } from '../../core/ApolloClient.js';
 
 declare module '../../core/ApolloClient.js' {
@@ -12,9 +12,11 @@ declare module '../../core/ApolloClient.js' {
 
 const suspenseCacheSymbol = Symbol.for('apollo.suspenseCache');
 
-export function getSuspenseCache(client: ApolloClient<object> & {
-  [suspenseCacheSymbol]?: SuspenseCache;
-}) {
+export function getSuspenseCache(
+  client: ApolloClient<object> & {
+    [suspenseCacheSymbol]?: SuspenseCache;
+  }
+) {
   if (!client[suspenseCacheSymbol]) {
     client[suspenseCacheSymbol] = new SuspenseCache(
       client.defaultOptions.react?.suspense
