+++ src/react/index.ts
@@ -10,17 +10,7 @@ export {
 
 export * from './hooks/index.js';
 // TODO: remove export with release 3.8
-export class SuspenseCache {
-  constructor(){
-    // throwing an error here instead of using invariant - we do not want this error
-    // message to be link-ified, but to directly show up as bold as possible
-    throw new Error(
-      'It is no longer necessary to create a `SuspenseCache` instance and pass it into the `ApolloProvider`.\n' + 
-      'Please remove this code from your application. \n\n' + 
-      'This export will be removed with the final 3.8 release.'
-    )
-  }
-}
+export { SuspenseCache } from './cache/index.js';
 
 export {
   DocumentType,
