+++ src/testing/matchers/toHaveSuspenseCacheEntryUsing.ts
@@ -3,7 +3,7 @@ import type { DocumentNode } from 'graphql';
 import type { OperationVariables } from '../../core/index.js';
 import { ApolloClient } from '../../core/index.js';
 import { canonicalStringify } from '../../cache/index.js';
-import { getSuspenseCache } from '../../react/hooks/getSuspenseCache.js';
+import { getSuspenseCache } from '../../react/cache/index.js';
 import type { CacheKey } from '../../react/cache/types.js';
 
 export const toHaveSuspenseCacheEntryUsing: MatcherFunction<
